requirements.txt
discord.py
subprocess
python-dotenv